Constrain examples to max width
Otherwise they expand to reveal the full width of the text content, which gives you a lot of horizontal scrolling, especially with examples with a lot of data on a single row.

A surrounding div has a hard-coded `style="display: inline-block"` definition which makes examples behave this way. No idea why it is there.